chore: remove old settings sidebar
